refactor v2 trace reduce fold
